[5043] * Fixed: prevent creature's auras duplication at creature respawn. * Foxed: use expected high guid for dynamic object instead explicit outdated value. * Make some compliers more happy with using high level options for warning checks.

+++ dep/src/zthread/posix/FtimeStrategy.h
@@ -52,7 +52,7 @@ public:
   }
 
   inline unsigned long seconds() const {
-    return _value.time;
+    return (unsigned long)_value.time;
   }
 
   inline unsigned long milliseconds() const {
